:truck: Move queryTokens to live in seerComboBox dir
