+++ static/app/views/explore/components/seerComboBox/queryTokens.tsx
